requirements.txt
ansible>=9.0.0
pyyaml
pandas
numpy
matplotlib
pyzmq
tabulate